+++ app/services/live/risk_manager_service.rb
@@ -30,6 +30,13 @@ module Live
 
     def initialize(exit_engine: nil)
       @exit_engine = exit_engine
+      @algo_config = begin
+        AlgoConfig.fetch
+      rescue StandardError
+        {}
+      end
+      @paper_mode = @algo_config.dig(:paper_trading, :enabled) == true
+      @orders_gateway = @paper_mode ? Orders::GatewayPaper.new : Orders::GatewayLive.new
       @mutex = Mutex.new
       @running = false
       @thread = nil